Add a separate launcher for .NET projects
